Fix path comparison in specs
The specs were using `toMatch`, which is for RegEx patterns not equality
comparisons. It just happens that on POSIX systems the path separator is
`/` which has no special meaning in RegEx patterns, but on Windows it is
`\` which caused all the paths to fail as it tried to interpret many
characters as escape sequences. Switch to the correct `toBe` to check
equality.