Cleanup, round 2 (#81)
* Fix path comparison in specs

The specs were using `toMatch`, which is for RegEx patterns not equality
comparisons. It just happens that on POSIX systems the path separator is
`/` which has no special meaning in RegEx patterns, but on Windows it is
`\` which caused all the paths to fail as it tried to interpret many
characters as escape sequences. Switch to the correct `toBe` to check
equality.

* Update Travis CI configuration for Atom Beta v1.19.0

Atom Beta v1.19.0 brings with it a native C++ implementation that
requires a newer `libstdc++6` version.

* Use a more forgiving method of getting the line

Instead of pulling in the entire file's lines, use a method that will
only give us the one we are interested in, and handles invalid lines
better.